Add a setup.py for building third with p2exe.
On a Windows machine which has all the prerequisites installed, calling
python.exe setup.py py2exe in the top-level directory of the source
distribution will build the application into an executable.

The prerequisites are:
 * Python
 * GTK+
 * pygtk
 * pycairo
 * pygobject
 * py2exe

Remove the win32 zipfile from the Makefile, as the final zip will need
to be packaged under Windows, after doing running the py2exe run.  Note
that for the win32 build to be workable, we will need to copy certain GTK
library files into the dist directory and include those files in the
zip.  This will need to be arranged to happen after the py2exe run.