Tweak py2exe options.
After much trial-and-error I seem to have stumbled upon the combination
of options which will actually create a third which can run on a clean
Windows machine.

The trickiest one was finding that I needed to deliberately exclude
DNSAPI.dll.  This file is included by py2exe's dependency detector, but
the symbols don't necessarily match on the destination machine.