Redirect stderr to a logfile in the user's app directory.
The original motivation was that py2exe's version of the app would try
to drop stderr messages out to third.exe.log in the program directory,
which was fine on winxp but failed due to permissions issues on win7.
Rather than bugger around trying to solve the permissions issues, I just
had the messages go to the user's private third directory, and having
done so I felt that this would be a reasonable arrangement for linux
too.