Get the basic layout working with composite widgets.
I have used a subclass of TableRow to represent the composite
arrangement of a dice button with its companion counter field.  The idea
is for the button to be capable of controlling the counter; clicking
on the button should advance the counter by one each time.  I haven't
yet built any of the listeners to make this work; this commit just
establishes the layout.