Make sure adb server is running in android-debug build target.
If the adb server is not running when you call `adb install`, it doesn't
work.  It does automatically start up the server, but it doesn't resume
the install task.  The `adb start-server` command will start up the
server if it isn't running but will harmlessly exit if it is already
running.