Implement rename/delete of presets.
This includes some behind-the-scenes work on how the ThirdActivity
stores its config.  Previously, the config variable was blown away and
rebuilt from the counters any time one of the counters was updated.  In
order to keep track of the last selected preset's id and name, I needed
to make the config persist, and to do that I had to break a nasty loop
condition.

The loop condition was triggered when I set the config.  This would
populate each of the counters with the value from the config, which
would in turn trigger the event to update the config from the counters!
I worked around this by setting a special flag called ConfigImmutable at
the start of updateFromConfig() and unsetting it at the end.  This way
the code knows to blithely ignore any updates *to* the config while it
is busy updating *from* the config.