Refactor the Counter classes to make room for the dx.
The Counter base class is now abstract and only knows how to interact
with the EditText of the counter itself; what goes into the left-hand
cell and how the row is laid out is left entirely up to the discretion
of inheriting classes.

The layout and button code has been shifted into a new ButtonCounter
class, which is instantiated directly to produce the multiplier and
modifier buttons.  The DiceButton class is now a subclass of
ButtonCounter.  In this way, the DxCounter will be able to inherit from
Counter with a totally different setup to the other counter types.