Add the dx.
The dx is a special part of the dice configuration which can have any
number of sides the user wishes.  This allows you to roll "dice" which
would be impossible to construct as a physical object, like a d3 or d27.

This commit also includes a lot of frobbing of the layout of the
counters to get them to behave on the HTC Hero.  I found that no amount
of pleading with attributes like "maxWidth" and "maxEms" could get the
counters to be any smaller than about 7 digits wide.  Setting the
layout_width attribute to a constant width did do the trick, however.