Fix buggy behavior on screen flip.
I found out to my chagrin that when you switch the phone's orientation,
the entire activity gets binned and re-created.  Normally, any view with
an id attribute has its state saved into a Bundle, which is then
restored when the view comes back up.  Unfortunately, this works about
as well as a porcelain cricket bat when multiple view have the same id;
as they do in my counters because I have reused the layout.

The result was that when you flip third into landscape, it not only
looks horrible but it sets ALL of the counters to 1.  Not helpful.

I fixed this by adding a screenOrientation attribute in the manifest, so
the orientation will stay in portrait, but that didn't stop android from
destroying and recreating my activity, so I still had that saved state
problem.

This was fixed by implementing my own onSaveInstanceState() and
onRestoreInstanceState() methods, which take care of saving and
restoring the value of the counters.