Add dependency on Android Support Library.
This allows us to support a wider range of API levels while still having
a standard-ish Action Bar.

I still need to convert the existing 'menu' into a modern toolbar-based
ActionBar.

The minimum API level is bumped from 3 to 9 (the lowest version
supported by the appcompat library), and the target API level is bumped
to 25.