Major update to persistent storage model.
Instead of using an SQLite database, which is patently overkill for the
amount of data this app stores, I am now serialising all of the app data
into a JSON string and storing that string as a SharedPreference.

The JSON string is an array of profile objects, and the 'presets' member
of each profile object is an array of preset objects.

Because this update required me to modify large portions of the code,
I've also taken this opportunity to clean up a lot of Lint issues, fix a
few misbehaviours and take greater care with user input checking.

This commit is a WIP, the use of preset includes does not appear to be
working at the moment.