Re-implement delete profile as an AlertDialog.
This fits in better with the standard UI than my giant full-screen
activities, but it sure was a pain to write.  The Android Dialog API is
a coder's nightmare of batteries-not-included nonsense.

Despite the ridiculous amount of boilerplate code I had to put in to
conjure up a simple confirmation dialog, this is still a win in terms of
lines of code once I took out the activity class and layout XML for the
old stuff.