Make the primary result much more prominent.
The most recent result is now shown centered in the right-hand pane, in
a huge text size, and the four results prior to it are shown separately
in a row below from left to right.

The previous layout was not visually obvious enough about displaying the
latest roll result, and the right-to-left ordering of result history was
unintuitive.