Require confirmation before deleting an included preset.
Also, when deleting a preset, go through existing presets and remove any
inclusions that point to the deleted preset.  This will happen in the
fullness of time anyway, as saving and reloading the profiles will
discard any include references that do not point to a valid preset, but
it's better all round to do it immediately.