Add initial code to support dice roll triggers.
The triggers mechanism is intended to allow for more complex roll
algorithms like "exploding" dice (rolling again on certain results, and
adding the new result) and "(dis)advantage" (rolling again and taking
the lesser/greater of the two rolls).

I have written up the first pass of logic for the triggers, but not any
UI to add or remove them from roll configurations in the app.