Save/restore instance state with current configuration as JSON.
This means that if the activity is killed and restarted by the OS, it
will retain the entire roll configuration, including the name,
inclusions and triggers.  Previously we just put the current value of
the major inputs into saved state, which didn't really make sense, since
the superclass automatically saves and restores the content of all the
inputs in the layout.  The job of my subclass methods is to make sure
that my custom variables (like mConfig) are correctly re-populated.