Use a view pager instead of a view flipper with radios.
In the phone layout, in order to switch between showing the profile/presets and the results details in the right-hand pane, I had previously used a ViewFlipper controlled by a pair of radio buttons.

That's an ugly and outdated approach.  The modern way to achieve this is
to make each of the switchable views into Fragments, and to switch
between them using a ViewPager, which is much more visually intuitive
and, as an added bonus, supports swipe gestures.

Further, now that the views are set up as Fragments, I can produce an
alternate layout for tablets that includes both fragments side-by-side,
with no need to switch between them.