Fix coding error in resolving inclusions.
For some daft reason I thought that it would make sense to use a
SparseIntArray to hold the relationships between presets and their
inclusions.  The problem is that when you have multiple inclusions for
one preset, they clobber each other, because SparseIntArray is a map.

What I actually needed here was either a list of (int, int) tuples, or a
map of int -> a list of ints.  I went with a list of (int, int) tuples
for simplicity.