Allow for negative dice counts.
When a dice count is negative, it means to roll the (absolute value)
number of dice, and subtract each result from the total instead of
adding each result.

The logic was mostly already set up to handle this correctly, although I
did have to correct one oversight in ThirdActivity.rollDie().  Apart
from that, it was just a matter of amending the inputType on the counter
inputs from "number" to "numberSigned".