Fix table layout of dice counter UI.
Turns out that I had been doing this wrong the whole time.
Instantiating a class that extends from TableRow, and inflating an XML
layout into it which has a TableRow root node actually yields a TableRow
inside another TableRow.

The correct approach for a situation like this is to use a <merge> node
as the root node of the XML, which actually collapses to be no node at
all once it is inflated.

Also, add a little "d" label in front of the dX sides input, to
hopefully make it more clear what these inputs do.