Add icons for minimum and maximum results.
Alongside the primary result text, there is a thumbs-down icon to the
left, and a thumbs-up icon to the right.  Both icons are initially
invisible.

On each roll, if the result matches the lowest possible result of the
roll, we show the thumbs-down icon.  Likewise, if it matches the highest
possible roll, we show the thumbs-up icon.  The intention is to draw the
user's attention to results that might trigger some special rule
(critical success or critical failure).